refactor: move files to new folder and rename folder
