+++ src/modules/Loads/LoadDetails/LoadDetails.module.scss
@@ -53,7 +53,7 @@
     padding: 0.5rem 0;
     font-weight: 700;
     &:hover {
-      background-color: $secondary-hover-text-color;
+      background-color: $secondary-hover-button-color;
     }
   }
 }
