+++ src/modules/Loads/LoadsList/components/AddLoad/AddLoad.tsx
@@ -8,7 +8,8 @@ import AlertDialog from '../../../../../common/Dialog/AlertDialog';
 import { useAppSelector, useAppDispatch } from '../../../../../store/hooks';
 import { closeModal, openDialog, closeDialog } from '../../../../../store/reducers/modalSlice';
 import { addLoadSchema, AddLoadValues } from '../../../../../utils/schemas/addLoadSchema';
-import { AddLoadData, addLoad } from '../../../../../utils/api/supabase/load';
+import { AddLoadData } from '../../../../../utils/api/supabase/types';
+import { addLoad } from '../../../../../utils/api/supabase/addLoad';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { PlaceAndDateSearch } from './PlaceAndDateSearch';
 import { CargoDetails } from './CargoDetails';
