+++ src/modules/Loads/LoadsList/components/AddLoad/CargoDetails.module.scss
