+++ src/modules/Loads/LoadsList/components/AddLoad/CargoDetails.tsx
