+++ src/modules/Loads/LoadsList/components/AddLoad/PaymentDetails.module.scss
