+++ src/modules/Loads/LoadsList/components/AddLoad/VehicleDetails.module.scss
