+++ src/modules/Loads/LoadsList/components/AddLoad/VehicleDetails.tsx
