+++ src/modules/Loads/LoadsList/loadData.ts
@@ -9,3 +9,5 @@ export const sortData = [
 ];
 
 export const currencies = ['EUR', 'PLN'];
+
+export const noFilterTab = 'all-loads';
